present  changes
