+++ src/main/resources/static/index.html
@@ -18,22 +18,59 @@
         button {
             padding: 10px 20px;
             font-size: 16px;
+            background-color: #616161
         }
+        .navigation {
+    text-align: center;
+    margin-top: 20px;
+}
+
+.navigation a {
+    text-decoration: none;
+    color: green; /* Dark gray text */
+    margin: 0 10px;
+    font-size: 16px;
+}
+
+.navigation a:hover {
+    color: #4CAF50; /* Green text on hover */
+}
     </style>
 </head>
 <body>
-<h1>Text Translator</h1>
-<form id="translateForm">
-    <textarea id="textInput" rows="10" cols="50" placeholder="Enter text to translate..."></textarea><br>
-    <button type="button" onclick="translateText()">Translate</button>
-</form>
-<h2>Translated Text</h2>
-<p id="translatedText"></p>
+<div class="container">
+    <h1>Text Translator</h1>
+    <form id="translateForm">
+        <textarea id="textInput" rows="10" placeholder="Enter text to translate..."></textarea><br>
+        <select id="languageSelect">
+            <option value="select">Select</option>
+            <option value="de">German</option>
+            <option value="fr">French</option>
+            <option value="es">Spanish</option>
+            <option value="it">Italian</option>
+            <option value="ru">Russian</option>
+            <option value="ar">Arabic</option>
+            <option value="hi">Hindi</option>
+            <option value="hu">Hungarian</option>
+            <option value="el">Greek</option>
+            <option value="ga">Irish</option>
+            <option value="zh">Chinese</option>
+            <option value="sv">Swedish</option>
+        </select>
+        <button type="button" onclick="translateText()">Translate</button>
+    </form>
+    <h2>Translated Text</h2>
+    <p id="translatedText" ></p>
+</div>
+</div>
+<div class="navigation">
+    <a href="http://localhost:8080/web Page.html">TranslationWebPage</a>
+</div>
 
 <script>
     async function translateText() {
         const textInput = document.getElementById('textInput').value;
-        const targetLang = 'es'; // Change this to the target language code as needed
+        const targetLang = document.getElementById('languageSelect').value;
         const response = await fetch('/translate-text', {
             method: 'POST',
             headers: {
